- added mtevent update and integrity testing to main update
