+++ api/tests/test_counter.py
@@ -5,6 +5,8 @@ import io
 import pandas as pd
 import base
 import json
+import numpy as np
+import datetime as dt
 
 from base.models import Position, ShipmentArrivalBerth
 from base.db import session
@@ -274,3 +276,54 @@ def test_counter_sorting(app):
 
         for c in counter_df.commodity.unique():
             assert list(counter_df[counter_df.commodity==c].date) == list(counter_df[counter_df.commodity==c].date.sort_values(ascending=False))
+
+def test_counter_against_voyage(app):
+    with app.test_client() as test_client:
+
+        response = test_client.get('/v0/counter')
+        assert response.status_code == 200
+        data = response.json["data"]
+        counter_df = pd.DataFrame(data).sort_values(['date'], ascending=False)
+
+        response = test_client.get('/v0/overland')
+        assert response.status_code == 200
+        data = response.json["data"]
+        pipeline_df = pd.DataFrame(data).sort_values(['date'], ascending=False)
+
+        response = test_client.get('/v0/voyage')
+        assert response.status_code == 200
+        data = response.json["data"]
+        voyage_df = pd.DataFrame(data)
+
+        counter2 = pd.concat([
+            voyage_df.loc[(voyage_df.arrival_date_utc>='2022-02-24')&(voyage_df.departure_iso2=='RU')][['destination_region', 'commodity_group','value_eur']],
+            pipeline_df.loc[(pipeline_df.date >= '2022-02-24') &
+                            (pipeline_df.departure_iso2.isin(['TR','RU','BY']))][['destination_region', 'commodity_group', 'value_eur']]]) \
+        .groupby(['destination_region', 'commodity_group']) \
+        .agg(value_eur=('value_eur', lambda x: np.nansum(x)/1e9))
+
+        counter1 = counter_df.groupby(['destination_region', 'commodity_group']) \
+        .agg(value_eur=('value_eur', lambda x: np.nansum(x)/1e9))
+
+        counter1==counter2
+
+def test_pricing_gt0(app):
+    with app.test_client() as test_client:
+        response = test_client.get('/v0/counter')
+        assert response.status_code == 200
+        data = response.json["data"]
+        counter_df = pd.DataFrame(data).sort_values(['date'], ascending=False)
+
+        response = test_client.get('/v0/overland')
+        assert response.status_code == 200
+        data = response.json["data"]
+        pipeline_df = pd.DataFrame(data).sort_values(['date'], ascending=False)
+
+
+        pipeline_notgas = pipeline_df.loc[pipeline_df.commodity != 'natural_gas']
+
+        assert pd.to_datetime(pipeline_df.date).max() > dt.date.today() - dt.timedelta(days=3)
+        assert all(pipeline_df.value_eur > 0)
+
+        assert pd.to_datetime(pipeline_notgas.date).max() > dt.date.today() - dt.timedelta(days=3)
+        assert all(pipeline_notgas.value_eur > 0)
